requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
joblib==1.5.1
motor==3.7.1
nltk==3.9.1
numpy==2.2.6
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
regex==2024.11.6
scikit-learn==1.7.1
scipy==1.15.3
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
